+++ module/apps/scene-app.mjs
@@ -21,7 +21,7 @@ export class MistSceneApp extends HandlebarsApplicationMixin(ApplicationV2) {
         tag: 'div',
         window: {
             frame: true,
-            title: 'Szene Tags',
+            title: 'Scene Tags',
             icon: 'fa-solid fa-book-atlas',
             positioned: true,
             resizable: true