+++ module/data/_module.mjs
@@ -12,3 +12,4 @@ export {default as MistEngineItemThemeBook} from "./item-themebook.mjs";
 export {default as MistEngineItemBackpack} from "./item-backpack.mjs";
 export {default as MistEngineSceneData} from "./item-scene-data.mjs";
 export {default as MistEngineActorFellowshipThemecard} from "./actor-fellowship-themecard.mjs"
+export {default as MistEngineQuintessence} from "./item-quintessence.mjs";