+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+# 0.2.6
+- quintessences are now items, hover over the quintessence name and you will see the description for easier referencing
+- added markiers for statuses
+- moved tags & status to the right side of the sheets
+- bug fixes here and there
+- Ctrl-T opens the scene tag window
+  
 # 0.2.5
 - fixed scrolling issues in NPC and character sheets